requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
psycopg2==2.8.6
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.24.2
